Create file, config file, etc.

+++ src/main.rs
@@ -1,38 +1,42 @@
 mod args;
 mod functions;
+mod read_config;
 
 use args::TodoArgs;
 use clap::Parser;
 use functions::create_file::create_file;
 use functions::delete_file::delete_file;
+use functions::list_files::list_files;
 use functions::open_file::open_file;
+use read_config::{config_exists, read_config};
 
 /// TODO
-/// Installation
-/// Ask for directory, if none provided use default directory
-/// Using
-/// Provide list of files to select
 /// Escape to go back
-///https://crates.io/crates/aarty
 
 fn main() {
-    let args = TodoArgs::parse();
+    let config = read_config();
+    let args_len = std::env::args().count() - 1;
 
-    // Clears the terminal
-    print!("\x1B[2J");
+    if config_exists() && args_len == 0 {
+        list_files(config.directory);
+    } else {
+        let args: TodoArgs = TodoArgs::parse();
+        // Clears the terminal
+        print!("\x1B[2J");
 
-    //Parses the arguments and calls the appropriate function
-    match args.first_arg.as_str() {
-        "create" => {
-            create_file(args.second_arg.as_str()).expect("TODO: panic message");
-            open_file(args.second_arg.as_str()).expect("TODO: panic message");
+        //Parses the arguments and calls the appropriate function
+        match args.first_arg.as_str() {
+            "create" => {
+                create_file(args.second_arg.as_str());
+                open_file(args.second_arg.as_str()).expect("TODO: panic message");
+            }
+            "open" => {
+                open_file(args.second_arg.as_str()).expect("TODO: panic message");
+            }
+            "delete" => {
+                delete_file(args.second_arg.as_str()).expect("TODO: panic message");
+            }
+            _ => {}
         }
-        "open" => {
-            open_file(args.second_arg.as_str()).expect("TODO: panic message");
-        }
-        "delete" => {
-            delete_file(args.second_arg.as_str()).expect("TODO: panic message");
-        }
-        _ => {}
     }
 }
